requirements.txt
streamlit
numpy
tensorflow-cpu
Pillow